requirements.txt
requests>=2.31.0
numpy>=1.24.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10
